Moved the GUI class to the gui package <organisation
